backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
celery==5.3.4
pika==1.3.2
pydantic==2.5.1
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
python-socketio==5.10.0
socketio==0.2.1
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
limits==3.10.1
slowapi==0.1.9
email-validator==2.1.0.post1
PyYAML==6.0.2
langchain==0.1.20
llama-cpp-python==0.2.79
aiohttp==3.9.5
yara-python==4.5.4
aio-pika>=8.0.0
asyncio
yarl
